requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
